refactor: split AssemblyDefinitionConsoleErrorValidationService into 500-line-clean files
Extract Class: move the AssemblyDefinitionConsoleError/Result DTOs into
their own files and move the pure CreateFailureMessage formatting into
AssemblyDefinitionConsoleErrorMessageFormatter, separating message
formatting from Console error detection. No behavior change; the
characterization tests added in the previous commit stay green.
