+++ Packages/src/Editor/FirstPartyTools/Compile/AssemblyDefinitionConsoleError.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9effbf84b0c47808273b67f44ee60f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
