+++ Packages/src/Editor/FirstPartyTools/Compile/AssemblyDefinitionConsoleErrorMessageFormatter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68e3d743817348eeb998dcf9327f9df7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
