+++ Packages/src/Editor/FirstPartyTools/Compile/AssemblyDefinitionConsoleErrorResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5242f82585684c609ab4fc7c2ee1c78f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
